requirements.txt
pdfminer.six==20231228
PyMuPDF==1.24.10
easyocr==1.7.2
fastapi==0.115.0
uvicorn==0.30.6

